Fixes for 64-bit AIX.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/1556



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@8170 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/canittst.cpp
@@ -104,7 +104,7 @@ void CanonicalIteratorTest::TestExhaustive() {
         }
         
         if (!gotSource || !gotDecomp || !gotComp) {
-            errln("FAIL CanonicalIterator: " + s + i);
+            errln("FAIL CanonicalIterator: " + s + (int)i);
         }
     }
 }
